+++ app/src/main/java/com/wanderreads/ebook/data/local/AppDatabase.kt
@@ -15,7 +15,7 @@ import com.wanderreads.ebook.data.local.entity.RecordEntity
  */
 @Database(
     entities = [BookEntity::class, RecordEntity::class],
-    version = 6,
+    version = 7,
     exportSchema = false
 )
 abstract class AppDatabase : RoomDatabase() {