+++ app/src/main/java/com/wanderreads/ebook/domain/model/BookType.kt
@@ -8,5 +8,6 @@ enum class BookType {
     PDF,
     TXT,
     MD,
+    WORD,
     UNKNOWN
 }