+++ app/src/main/java/com/wanderreads/ebook/ui/bookshelf/BookshelfScreen.kt
@@ -634,6 +634,7 @@ fun BookListItem(
                         BookType.PDF -> MaterialTheme.colorScheme.secondaryContainer
                         BookType.TXT -> MaterialTheme.colorScheme.surfaceVariant
                         BookType.MD -> MaterialTheme.colorScheme.surfaceVariant
+                        BookType.WORD -> MaterialTheme.colorScheme.secondaryContainer
                         BookType.UNKNOWN -> MaterialTheme.colorScheme.tertiaryContainer
                     }
                     
@@ -766,6 +767,7 @@ fun BookCover(
                 BookType.PDF -> MaterialTheme.colorScheme.secondaryContainer
                 BookType.TXT -> MaterialTheme.colorScheme.surfaceVariant
                 BookType.MD -> MaterialTheme.colorScheme.surfaceVariant
+                BookType.WORD -> MaterialTheme.colorScheme.secondaryContainer
                 BookType.UNKNOWN -> MaterialTheme.colorScheme.tertiaryContainer
             }