+++ app/src/main/java/com/wanderreads/ebook/util/reader/BookReaderEngine.kt
@@ -146,6 +146,7 @@ interface BookReaderEngine {
                 BookType.PDF -> BookFormat.PDF
                 BookType.TXT -> BookFormat.TXT
                 BookType.MD -> BookFormat.TXT // 将MD格式当作TXT处理
+                BookType.WORD -> BookFormat.DOCX // 将WORD格式当作DOCX处理
                 BookType.UNKNOWN -> BookFormat.UNKNOWN
             }
             return ReaderEngineFactory.createEngine(context, format)